Folder structure and code refactor
