--- src/Display.css
@@ -1,101 +1,0 @@
-/* .cardsContainer {
-	display: flex;
-	flex-direction: row;
-	align-items: center;
-	justify-content: center;
-	flex-wrap: wrap;
-	width: 95%;
-	margin: 0 auto;
-}
-
-.cardsContainer {
-	flex: 1;
-}
-
-.card:nth-child(odd) {
-	border-right: 1px solid grey;
-}
-
-.card:nth-child(even) {
-	border-right: 1px solid white;
-	border-left: 1px solid grey;
-	margin-left: -1px;
-}
-
-.author {
-	display: flex;
-	flex-wrap: wrap;
-	flex-direction: row;
-	justify-content: space-between;
-}
-
-.content {
-	padding-top: 0;
-	width: 58%;
-	padding: 5px;
-}
-
-.card {
-	display: flex;
-	flex-direction: row;
-	width: 49%;
-	padding: 5px;
-	margin: 10px 0;
-	border-bottom: 1px solid grey;
-}
-
-.image {
-	padding-left: 1rem;
-	width: 40%;
-	margin-bottom: 0;
-}
-
-.cardsContainer img {
-	max-width: 300px;
-	width: 100%;
-}
-
-@media screen and (max-width: 1200px) {
-	.cardsContainer {
-		display: flex;
-		flex-direction: column;
-	}
-	.card:nth-child(odd) {
-		border-right: 0;
-	}
-	.card {
-		width: 90%;
-	}
-}
-
-@media screen and (max-width: 600px) {
-	.card .image {
-		order: -2;
-		width: 80%;
-		padding-left: 0;
-		margin-top: 2rem;
-	}
-	.card:nth-child(even) {
-	margin-left: inherit;
-	}
-	.card {
-		display: flex;
-		flex-direction: column;
-		align-items: center;
-		padding: 5px;
-		width: 80%;
-		margin: 0 auto;
-	}
-	.content {
-		width: 80%;
-	}
-	.cardsContainer {
-		justify-content: center;
-		align-items: center;
-	}
-} */
-
-.my-car-image{
-  height: 100px;
-  width: 100px;
-}
